INC-104 feat: select controlled & refactoring user profile settings page
