+++ src/components/controlled-select/index.ts
@@ -1,0 +1,1 @@
+export { ControlledSelect } from './controlled-select'
