--- src/components/countries-select/index.ts
@@ -1,1 +1,0 @@
-export { CountriesSelect } from './countries-select'
