+++ src/components/index.ts
@@ -9,3 +9,4 @@ export * from './avatar'
 export * from './calendar'
 export * from './user-profile'
 export * from './devices'
+export * from './controlled-select'
