+++ src/ui/custom-select/custom-select.types.ts
@@ -1,1 +1,10 @@
 export type SelectValue = { value: string; label: string }
+export type CustomSelectProps = {
+  value?: string
+  label?: string
+  options?: SelectValue[]
+  setClear?: () => void
+  isClearable?: boolean
+  placeholder?: string
+  onChange?: (value: string | undefined) => void
+}
